Update Makefile to fetch and build kernel source
The kernel module targets the Linux kernel used by Debian Buster. The
changes to the Makefile provide two new targets. The ksource target
fetches the source of the kernel from Debian repositories and applies
the Debian patches. The kbuild target builds the kernel.

The kernel module can then be built on a different host using:
$ make kbuild
$ make KDIR=./usr/src/linux-source-4.19

For code completion support this should be run as:
$ make kbuild
$ bear make KDIR=./usr/src/linux-source-4.19