Add logging library to client application
This commit also introduces a number of debug statements to track
control flow. This can be used to debug errors in the front-end or
back-end of the sc-implant project.

Logging can be enabled by compiling the project with:
$ make CFLAGS="IMLOG_LEVEL=0" sc-client
Note that the value for IMLOG_LEVEL corresponds to the level specified
in client/include/debug.h.