Add argument handler to client for privilege
This is the first work on communicating with the implant to fulfil some
request. The client program can now be used to modify the privilege of
any running process on the system. The impant does not seem to have the
feature to hande deescalating yet though the interface to the client
supports that.