Add process hiding/unhiding capability in the client
With the API implemented for process hiding and unhiding, the client
now has the capability of writing to the virtual device to execute process
modifications, including process hiding and unhiding. This implementation
of process modification draws from the implementation of privilege
modification (escalation/de-escalation). It mirrors the capabilities
present in the kernel module interface, including the commands and
arguments parsed for process modification.