Add interactive shell
This is a major refactor of the frontend client for sc-implant. The new
client uses the second iteration of the API. In addition to the new API,
a new modular structure for the project is included in the changes.
Subcommands are considered as "modules" and remain isolated from the
rest of the code base while following a very simple but strict
interface. This ensures that extension of the client program is easy so
that it can be customised to meet specific needs.

The interactive shell is triggered when there are no commands provided.
Where possible, it provides live feedback on the behaviour and success
of the dropper and implant while providing help where needed.