Add client tests to Gitlab CI
This commit adds tests for the client in the CI/CD pipeline. The tests
are all in a new stage of the pipeline and all extend a testing
template. The tests all check if the output from the client to the
device were as expected. This utilises the compile-time generated path
for the device. There is also a test to verify the output from the help
command.

Resolves: [ch287]