Fix structure of client tests
This commit refactors the test suite for the client. The restructure
aims to make it easier for other developers to add new tests. This is
achieved that by separating the scripts from the `.gitlab-ci.yml` file
into separate `.sh` and output files (`.stdout`, `.comm`). These are
all stored in `client/test/` under corresponding module directories
such as `exec` and `hide`.