Add test coverage reporting (Gitlab CI)
This commit adds test coverage reporting to the project. The reporting
is provided by CodeCov and test results are uploaded on every commit.
Once uploaded, the results are collated and fed into a dashboard which
is accessible to anyone in the corresponding Gitlab project,
`team-sc/sc-implant`. To view the dashboard, sign in or create an
account with the same Gitlab account used for this project.

Resolves: [ch289]